requirements.txt
tensorflow==2.12.0
streamlit
numpy
pillow
scipy
opencv-python
matplotlib
psutil
